View cone and circle, and image fading to black added
I added a panel that fades to black as the mouse scroll wheel increases to a certain measurement or decreases to a certain specific negative measurement. Also I put in a circle and view cone item that tells the user where they're rotated to in the image as well as how focused they are on it.