"lock" z axis, fade panel when falcon moves away
Added a script that will apply a force opposite to the user's movement in the z axis. It doesn't lock the axis, but it's hard to move very far from it.

The image fading now works with the falcon. I just replaced scrollPos with the falcon's position in the z axis when the falcon is being used.