Integrate falcon with tutorial, update UI
In the state manager I added an integer called buttons which will keep track of the buttons pressed for the current frame. For some reason the falcon initializes with some buttons pressed, so i added a delay to the statemanager so it wouldnt start immediately.

When going through the tutorial some of the text was sticking outside of its containing panel, so i made the panel bigger.

Also added FalconServer.exe for using the falcon