bugfixes, ui update
Tags weren't being loaded properly from the build because of the way that the csv file is loaded, so I created a TextAsset to contain the contents of the csv and used a StringReader in place of the StreamReader.

The positions of the arrows and other UI elements were different between running the game in the editor and the build, so I put them back to where they were originally. My bad!